# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Daily Shop Ledger API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface User {
  id: number;
  username: string;
  /** @nullable */
  email?: string | null;
  language: string;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  message: string;
}

export interface SignupBody {
  username: string;
  password: string;
  /** @nullable */
  email?: string | null;
  language?: string;
}

export interface LoginBody {
  username: string;
  password: string;
}

export interface ForgotPasswordBody {
  email: string;
}

export interface ResetPasswordBody {
  token: string;
  newPassword: string;
}

export type EntryType = (typeof EntryType)[keyof typeof EntryType];

export const EntryType = {
  cash_in: "cash_in",
  cash_out: "cash_out",
} as const;

export type EntryPaymentMethod =
  (typeof EntryPaymentMethod)[keyof typeof EntryPaymentMethod];

export const EntryPaymentMethod = {
  cash: "cash",
  digital: "digital",
} as const;

export interface Entry {
  id: number;
  userId: number;
  type: EntryType;
  amount: number;
  /** @nullable */
  description?: string | null;
  paymentMethod: EntryPaymentMethod;
  /** @nullable */
  profit?: number | null;
  isCredit: boolean;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  deletedAt?: string | null;
  entryDate: string;
  createdAt: string;
  updatedAt: string;
}

export type CreateEntryBodyType =
  (typeof CreateEntryBodyType)[keyof typeof CreateEntryBodyType];

export const CreateEntryBodyType = {
  cash_in: "cash_in",
  cash_out: "cash_out",
} as const;

export type CreateEntryBodyPaymentMethod =
  (typeof CreateEntryBodyPaymentMethod)[keyof typeof CreateEntryBodyPaymentMethod];

export const CreateEntryBodyPaymentMethod = {
  cash: "cash",
  digital: "digital",
} as const;

export interface CreateEntryBody {
  type: CreateEntryBodyType;
  amount: number;
  /** @nullable */
  description?: string | null;
  paymentMethod: CreateEntryBodyPaymentMethod;
  isCredit?: boolean;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  entryDate?: string | null;
}

export type UpdateEntryBodyType =
  (typeof UpdateEntryBodyType)[keyof typeof UpdateEntryBodyType];

export const UpdateEntryBodyType = {
  cash_in: "cash_in",
  cash_out: "cash_out",
} as const;

export type UpdateEntryBodyPaymentMethod =
  (typeof UpdateEntryBodyPaymentMethod)[keyof typeof UpdateEntryBodyPaymentMethod];

export const UpdateEntryBodyPaymentMethod = {
  cash: "cash",
  digital: "digital",
} as const;

export interface UpdateEntryBody {
  type?: UpdateEntryBodyType;
  amount?: number;
  /** @nullable */
  description?: string | null;
  paymentMethod?: UpdateEntryBodyPaymentMethod;
  isCredit?: boolean;
  /** @nullable */
  customerName?: string | null;
  /** @nullable */
  entryDate?: string | null;
}

export interface UpdateProfitBody {
  /** @nullable */
  profit: number | null;
}

export type CreditType = (typeof CreditType)[keyof typeof CreditType];

export const CreditType = {
  given: "given",
  received: "received",
} as const;

export type CreditStatus = (typeof CreditStatus)[keyof typeof CreditStatus];

export const CreditStatus = {
  pending: "pending",
  paid: "paid",
} as const;

export interface Credit {
  id: number;
  userId: number;
  customerName: string;
  amount: number;
  /** @nullable */
  description?: string | null;
  type: CreditType;
  status: CreditStatus;
  /** @nullable */
  dueDate?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateCreditBodyType =
  (typeof CreateCreditBodyType)[keyof typeof CreateCreditBodyType];

export const CreateCreditBodyType = {
  given: "given",
  received: "received",
} as const;

export interface CreateCreditBody {
  customerName: string;
  amount: number;
  /** @nullable */
  description?: string | null;
  type: CreateCreditBodyType;
  /** @nullable */
  dueDate?: string | null;
}

export type UpdateCreditBodyStatus =
  (typeof UpdateCreditBodyStatus)[keyof typeof UpdateCreditBodyStatus];

export const UpdateCreditBodyStatus = {
  pending: "pending",
  paid: "paid",
} as const;

export interface UpdateCreditBody {
  status?: UpdateCreditBodyStatus;
  amount?: number;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  dueDate?: string | null;
}

export interface Customer {
  id: number;
  name: string;
  totalCredit: number;
}

export interface BalanceSummary {
  cashBalance: number;
  digitalBalance: number;
  totalBalance: number;
  totalCashIn: number;
  totalCashOut: number;
  totalProfit: number;
  totalCredit: number;
  todayEntries: number;
}

export interface EntriesReport {
  period: string;
  startDate: string;
  endDate: string;
  totalCashIn: number;
  totalCashOut: number;
  cashBalance: number;
  digitalBalance: number;
  entries: Entry[];
}

export interface ProfitReport {
  period: string;
  startDate: string;
  endDate: string;
  totalProfit: number;
  entriesWithProfit: Entry[];
}

export interface BackupData {
  exportedAt: string;
  user: User;
  entries: Entry[];
  credits: Credit[];
}

export type ListEntriesParams = {
  type?: ListEntriesType;
  payment_method?: ListEntriesPaymentMethod;
  date?: string;
  start_date?: string;
  end_date?: string;
  deleted?: boolean;
};

export type ListEntriesType =
  (typeof ListEntriesType)[keyof typeof ListEntriesType];

export const ListEntriesType = {
  cash_in: "cash_in",
  cash_out: "cash_out",
} as const;

export type ListEntriesPaymentMethod =
  (typeof ListEntriesPaymentMethod)[keyof typeof ListEntriesPaymentMethod];

export const ListEntriesPaymentMethod = {
  cash: "cash",
  digital: "digital",
} as const;

export type ListCreditsParams = {
  customer_name?: string;
  status?: ListCreditsStatus;
};

export type ListCreditsStatus =
  (typeof ListCreditsStatus)[keyof typeof ListCreditsStatus];

export const ListCreditsStatus = {
  pending: "pending",
  paid: "paid",
} as const;

export type ListCustomersParams = {
  q?: string;
};

export type GetEntriesReportParams = {
  period: GetEntriesReportPeriod;
  date?: string;
};

export type GetEntriesReportPeriod =
  (typeof GetEntriesReportPeriod)[keyof typeof GetEntriesReportPeriod];

export const GetEntriesReportPeriod = {
  daily: "daily",
  weekly: "weekly",
  monthly: "monthly",
} as const;

export type GetProfitReportParams = {
  period: GetProfitReportPeriod;
  date?: string;
};

export type GetProfitReportPeriod =
  (typeof GetProfitReportPeriod)[keyof typeof GetProfitReportPeriod];

export const GetProfitReportPeriod = {
  daily: "daily",
  weekly: "weekly",
  monthly: "monthly",
} as const;