# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Daily Shop Ledger API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user info
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordBody"
      responses:
        "200":
          description: Reset email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordBody"
      responses:
        "200":
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /credits:
    get:
      operationId: listCredits
      tags: [credits]
      summary: List all credit entries
      parameters:
        - name: customer_name
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, paid]
      responses:
        "200":
          description: List of credits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Credit"

    post:
      operationId: createCredit
      tags: [credits]
      summary: Create a credit entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCreditBody"
      responses:
        "201":
          description: Credit created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Credit"
  /credits/{id}:
    patch:
      operationId: updateCredit
      tags: [credits]
      summary: Update a credit entry (e.g., mark as paid)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCreditBody"
      responses:
        "200":
          description: Credit updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Credit"

    delete:
      operationId: deleteCredit
      tags: [credits]
      summary: Delete a credit entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: ["string", "null"]
        language:
          type: string
        createdAt:
          type: string
      required:
        - id
        - username
        - language
        - createdAt
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required:
        - user
        - message
    SignupBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        email:
          type: ["string", "null"]
        language:
          type: string
      required:
        - username
        - password
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    ForgotPasswordBody:
      type: object
      properties:
        email:
          type: string
      required:
        - email
    ResetPasswordBody:
      type: object
      properties:
        token:
          type: string
        newPassword:
          type: string
      required:
        - token
        - newPassword
...
    Credit:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        customerName:
          type: string
        amount:
          type: number
        description:
          type: ["string", "null"]
        type:
          type: string
          enum: [given, received]
        status:
          type: string
          enum: [pending, paid]
        dueDate:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - customerName
        - amount
        - type
        - status
        - createdAt
        - updatedAt
    CreateCreditBody:
      type: object
      properties:
        customerName:
          type: string
        amount:
          type: number
        description:
          type: ["string", "null"]
        type:
          type: string
          enum: [given, received]
        dueDate:
          type: ["string", "null"]
      required:
        - customerName
        - amount
        - type
    UpdateCreditBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, paid]
        amount:
          type: number
        description:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
...