# lib/api-zod/src/generated/types/signupBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Daily Shop Ledger API
 * OpenAPI spec version: 0.1.0
 */

export interface SignupBody {
  username: string;
  password: string;
  /** @nullable */
  email?: string | null;
  language?: string;
}